Add CreateEncryptRuleStatementUpdater
